+++ Tools/CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Boolean and float parsing settings customization
 - Yandex documents support
 
-## [0.1.0] - 2022-04-20
+## [0.1.1] - 06-05-2023
+### Updated
+- .asmdef
+- Minor naming
+### Added
+- CodeFileContent wrapper class
+- ContainerScriptableObject class
+
+## [0.1.0] - 20-04-2023
 ### Added
 - Basic functionality